--- packages/DataTable/src/components/Events/Events.js
@@ -1,5 +1,0 @@
-export default function Events() {
-  return null;
-}
-
-Events.displayName = "DataTable.Events";
