--- packages/DataTable/src/components/Events/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Events";
