+++ packages/DataTable/src/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Table";
+export { default } from "./DataTable";
